reorganizeClasses
