+++ src/servers/ZoneServer/commands/dev.ts
@@ -1,4 +1,4 @@
-import { ZoneClient as Client } from "../../ZoneServer/zoneclient";
+import { ZoneClient as Client } from "../classes/zoneclient";
 import { generateRandomGuid } from "../../../utils/utils";
 import { ZoneServer } from "../zoneserver";
