+++ src/servers/ZoneServer2016/classes/character.ts
@@ -1,4 +1,4 @@
-import { Character } from "servers/ZoneServer/character";
+import { Character } from "servers/ZoneServer/classes/character";
 
 export class Character2016 extends Character {
   resources: {
