+++ src/servers/ZoneServer2016/classes/zoneclient.ts
@@ -1,5 +1,5 @@
-import SOEClient from "../SoeServer/soeclient";
-import { ZoneClient } from "servers/ZoneServer/zoneclient";
+import SOEClient from "../../SoeServer/soeclient";
+import { ZoneClient } from "servers/ZoneServer/classes/zoneclient";
 import { Character2016 } from "./character";
 
 export class ZoneClient2016 extends ZoneClient {
